+++ Sources/UIModel.h
@@ -21,7 +21,7 @@ public:
 	explicit UIModel(const std::shared_ptr<VulkanCore> &vulkanCore);
 	virtual ~UIModel();
 
-	virtual void RecordCommand(VkCommandBuffer commandBuffer, uint32_t currentFrame) override;
+	virtual void RecordCommand(VkCommandBuffer commandBuffer, size_t currentFrame) override;
 	virtual uint32_t GetOrder() override { return 2000; }
 
 	template<typename TPanel, typename... TArgs>
